+++ client/src/App.js
@@ -2,15 +2,15 @@ import React, { Suspense, lazy } from 'react';
 import { Route, Switch } from "react-router-dom";
 import Auth from "./hooks/auth";
 
-const Landing = lazy(() => import("./Landing/Landing"))
-const Login = lazy(() => import("./Login/Login"))
-const Register = lazy(() => import("./Register/Register"))
-const Nav = lazy(() => import("./Nav/Nav"))
-const Footer = lazy(() => import("./Footer/Footer"))
-const UploadProduct = lazy(() => import("./UploadProduct/UploadProduct"))
-const DetailProduct = lazy(() => import("./DetailProduct/DetailProduct"))
-const Cart = lazy(() => import("./Cart/Cart"))
-const History = lazy(() => import("./History/History"))
+const Landing = lazy(() => import("./components/Landing/Landing"))
+const Login = lazy(() => import("./components/Login/Login"))
+const Register = lazy(() => import("./components/Register/Register"))
+const Nav = lazy(() => import("./components/Nav/Nav"))
+const Footer = lazy(() => import("./components/Footer/Footer"))
+const UploadProduct = lazy(() => import("./components/UploadProduct/UploadProduct"))
+const DetailProduct = lazy(() => import("./components/DetailProduct/DetailProduct"))
+const Cart = lazy(() => import("./components/Cart/Cart"))
+const History = lazy(() => import("./components/History/History"))
 
 function App() {
   return (
