+++ client/src/components/Nav/Nav.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import LeftMenu from './Sections/LeftMenu';
 import RightMenu from './Sections/RightMenu';
 import { Drawer, Button, Icon } from 'antd';
-import './Sections/Navbar.css';
+import './Sections/Nav.css';
 
 function Nav() {
   const [visible, setVisible] = useState(false)
