+++ client/src/store/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import user from './user_reducer';
+import user from './user';
 
 const rootReducer = combineReducers({
     user,
